# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Yangi foydalanuvchini ro'yxatdan o'tkazish",
                "description": "register() — POST /register",
                "operationId": "858ecba1c675ee39e49f2b0c265443ff",
                "parameters": [
                    {
                        "name": "auth_type",
                        "in": "query",
                        "description": "Registratsiya turi: password yoki telegram",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "password",
                            "enum": [
                                "password",
                                "telegram"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password",
                                    "fname",
                                    "lname"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+998901234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "parol123"
                                    },
                                    "fname": {
                                        "type": "string",
                                        "example": "Ism"
                                    },
                                    "lname": {
                                        "type": "string",
                                        "example": "Familiya"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "telegram": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Muvaffaqiyatli ro'yxatdan o'tdi, token qaytdi"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Telefon raqam va parol bilan login",
                "description": "login() — POST /login",
                "operationId": "a99c7c5fa27adf14888e3da5b618cfe6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+998901234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "parol123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli login, JWT token qaytadi"
                    },
                    "401": {
                        "description": "Telefon raqam yoki parol noto'g'ri"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Tizimdan chiqish",
                "description": "logout() — POST /logout",
                "operationId": "9c173571cff516089ecbcb956c008fad",
                "responses": {
                    "200": {
                        "description": "Muvaffaqiyatli logout"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Hozirgi foydalanuvchi ma'lumotlari",
                "description": "me() — GET /get-me",
                "operationId": "19ada446963c7a334dc487633636b104",
                "responses": {
                    "200": {
                        "description": "Foydalanuvchi ma'lumotlari"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Tokenni yangilash",
                "description": "refresh() — POST /refresh",
                "operationId": "19989e3128942ce9212927fec5791464",
                "responses": {
                    "200": {
                        "description": "Yangi token qaytadi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/public/ads": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Barcha faol e'lonlar (ixtiyoriy filter)",
                "description": "category_id/subcategory_id yuborilmasa — barcha faol e'lonlar; yuborilsa — shu bo'yicha filter.",
                "operationId": "728b62ca4aa4bee130a611e83c0407cd",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "1–50, default 15",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Berilmasa barcha kategoriyalar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subcategory_id",
                        "in": "query",
                        "description": "Berilmasa filter yo'q",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginatsiyali e'lonlar"
                    },
                    "422": {
                        "description": "category_id/subcategory_id noto'g'ri"
                    }
                }
            }
        },
        "/public/ads/{id}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Bitta faol e'lon",
                "description": "ad() — GET /public/ads/{id}",
                "operationId": "3dedff3ec13a5e42171cd56f935d0232",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E'lon obyekti"
                    },
                    "404": {
                        "description": "Topilmadi yoki faol emas"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}